Implementado Cadastro de Alugueis #Fé

+++ src/FestasInfantis.WinApp/ModuloAluguel/Aluguel.cs
@@ -7,14 +7,14 @@ namespace FestasInfantis.WinApp.ModuloAluguel
     public class Aluguel : EntidadeBase
     {
         public bool Status { get; set; }
-        public double PorcentagemEntrada { get; set; }
+        public int PorcentagemEntrada { get; set; }
         public double PorcentagemDesconto { get; set; }
         public DateTime DataPagamento { get; set; }
         public Cliente Cliente { get; set; }
         public Tema Tema { get; set; }
         public Festa Festa { get; set; }
 
-        public Aluguel(double porcentagemEntrada, double porcentagemDesconto, Cliente cliente, Tema tema, Festa festa)
+        public Aluguel(int porcentagemEntrada, double porcentagemDesconto, Cliente cliente, Tema tema, Festa festa)
         {
             Status = false;
             PorcentagemEntrada = porcentagemEntrada;
